Fix issue with Local.isolate(future) when modifying two locals
